1.DAO层重构 2.中心钱包模块：转账方法与提现方法重构 3.删除以前测试的Item与Hello 4.更改判断支付密码bug 5.更改初始化钱包信息请求 6.更新sql

+++ src/main/java/com/superwallet/controller/CWalletController.java
@@ -58,7 +58,7 @@ public class CWalletController {
     @RequestMapping(value = "/cWallet/transferMoney", method = RequestMethod.POST)
     @ResponseBody
     @Transactional
-    public SuperResult transferMoney(String UID, int tokenType, int tokenAmount) {
+    public SuperResult transferMoney(String UID, int tokenType, double tokenAmount) {
         boolean result = cWalletService.transferMoney(UID, tokenType, tokenAmount);
         if (result) {
             return SuperResult.ok();
@@ -77,7 +77,7 @@ public class CWalletController {
     @RequestMapping(value = "/cWallet/withdraw", method = RequestMethod.POST)
     @ResponseBody
     @Transactional
-    public SuperResult withdraw(String UID, int tokenType, int tokenAmount, HttpServletRequest request) {
+    public SuperResult withdraw(String UID, int tokenType, double tokenAmount, HttpServletRequest request) {
         boolean result = cWalletService.withdraw(UID, tokenType, tokenAmount);
         if (result) return SuperResult.ok();
         return new SuperResult(CodeRepresentation.CODE_FAIL, CodeRepresentation.STATUS_0, null);
